dapp-demo/requirements.txt
requests==2.22.0
Django==2.2.1
fastecdsa==1.7.3
pysha3==1.0.2
six==1.12.0
hep-rest-api==1.0.0
